backend/core/orchestator/orchestrator.py
from typing import Dict, Any, List
from pydantic import BaseModel
import networkx as nx

class Entity(BaseModel):
    id: str
    name: str
    description: str
    superclass: str = None
    properties: Dict[str, Any] = {}
    neighbors: List[str] = []

class Orchestrator:
    def __init__(self):
        self.current_entity = None
        self.graph = nx.Graph()
        self.explored_nodes = set()
        self.considering_nodes = set()
        self.entity_cache = {}

    def set_current_entity(self, entity_id: str) -> None:
        """Set the current entity to explore."""
        self.current_entity = entity_id
        self.explored_nodes.add(entity_id)
        self.considering_nodes.clear()

    def get_entity_info(self, entity_id: str) -> Entity:
        """Get entity information from cache or fetch it."""
        if entity_id in self.entity_cache:
            return self.entity_cache[entity_id]
        
        # TODO: Implement Wikidata API call
        entity = Entity(
            id=entity_id,
            name=f"Entity {entity_id}",
            description="Entity description",
            superclass="Q1"  # Root entity
        )
        self.entity_cache[entity_id] = entity
        return entity

    def explore_neighbors(self, entity_id: str) -> List[str]:
        """Explore and return neighbors of the given entity."""
        entity = self.get_entity_info(entity_id)
        neighbors = []
        
        # TODO: Implement neighbor fetching from Wikidata
        # For now, add some dummy neighbors
        if entity_id == "Q1":
            neighbors = ["Q2", "Q3", "Q4"]
        
        for neighbor in neighbors:
            self.graph.add_edge(entity_id, neighbor)
            if neighbor not in self.explored_nodes:
                self.considering_nodes.add(neighbor)
        
        return neighbors

    def get_graph_data(self) -> Dict:
        """Get graph data for visualization."""
        nodes = []
        links = []
        
        for node in self.graph.nodes:
            color = "gray"
            border_color = "darkgray"
            
            if node == self.current_entity:
                color = "green"
            elif node in self.explored_nodes:
                color = "white"
                border_color = "black"
            elif node in self.considering_nodes:
                border_color = "yellow"
            
            nodes.append({
                "id": node,
                "color": color,
                "borderColor": border_color
            })
        
        for edge in self.graph.edges:
            links.append({
                "source": edge[0],
                "target": edge[1]
            })
        
        return {"nodes": nodes, "links": links}

    def process_query(self, query: str) -> Dict:
        """Process a natural language query."""
        # TODO: Implement query processing
        response = {
            "result": "Query processing not implemented",
            "explored_nodes": list(self.explored_nodes),
            "query_path": []
        }
        return response
