backend/webapp/server.py
import websockets

@app.get("/")
async def root():
    return {"message": "Welcome to RoboData API"}

@app.websocket("/ws")
async def websocket_endpoint(websocket: WebSocket):
    await websocket.accept()
    try:
        while True:
            data = await websocket.receive_text()
            # Process websocket messages
            response = await process_message(data)
            await websocket.send_text(response)
    except Exception as e:
        print(f"WebSocket error: {e}")
    finally:
        await websocket.close()

async def process_message(message: str) -> str:
    """Process incoming websocket message."""
    # TODO: Implement websocket message processing
    return "Message processing not implemented"