requirements.txt
# Core dependencies
fastapi
hypercorn
python-dotenv
pydantic>=2.0.0
requests>=2.31.0
aiohttp>=3.8.0
PyYAML>=6.0

# LLM providers
openai
google-generativeai>=0.3.0

# REST API client for Wikidata
wikibase-rest-api-client>=0.1.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: for visualization (future use)
# matplotlib>=3.7.0
# networkx>=3.1.0
python-multipart
aiosqlite
wikidata
kif